Not if/then/else similarities to other languages
Maybe it's just me, but early on I felt scared to use if/then/else to do conventional things like if FileIsOK then WriteToFile because the help text emphasised that we are dealing with a cool new expression construct not the lame if/then/else statement of lesser languages. Scott Wlaschin goes as far as to say use of if/then/else is strongly discouraged.

So my suggestion here is to say, "if/then/else in F# is the same as other languages in some ways (so relax) but different in others (so be alert).